backend/requirements.txt
fastapi
uvicorn
python-multipart
starlette
pydantic
PyMuPDF
python-docx
Pillow
pytesseract
motor
qdrant-client
sentence-transformers
python-dotenv
openai
tenacity
rapidfuzz
msal
amazon_transcribe
googlemaps
elevenlabs